requirements.txt
Django==1.9.2
django-extensions==1.6.1
djangorestframework==3.3.2
drf-extensions==0.2.8
requests==2.9.1
django-cors-headers==1.1.0
MySQL-python==1.2.5
jsonfield==1.0.3
pygments==2.1.3
